Refactor tab modules and encapsulate UI logic
