+++ public/js/tabs/finishing.js
@@ -1,5 +1,19 @@
+let initialized = false;
+
+function init() {
+  if (initialized) return;
+  initialized = true;
+}
+
 const finishingTab = {
   key: 'finishing',
+  init,
+  onActivate() {
+    init();
+  },
+  onRegister() {
+    init();
+  },
 };
 
 export default finishingTab;
