+++ public/js/tabs/warnings.js
@@ -1,5 +1,19 @@
+let initialized = false;
+
+function init() {
+  if (initialized) return;
+  initialized = true;
+}
+
 const warningsTab = {
   key: 'warnings',
+  init,
+  onActivate() {
+    init();
+  },
+  onRegister() {
+    init();
+  },
 };
 
 export default warningsTab;
